% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cow.R
\name{create_cow}
\alias{create_cow}
\title{Create Cohort Weights (COW) for Multi-Cohort Data}
\usage{
create_cow(
  data,
  cohort_id,
  target_cohort,
  baseline_covs,
  trunc_weights = TRUE,
  weight_threshold = NULL
)
}
\arguments{
\item{data}{A dataframe containing the multi-cohort data.}

\item{cohort_id}{A character string indicating the column name which specifies cohort labels in the `data`.}

\item{target_cohort}{A character or numeric value specifying which cohort is the target.}

\item{baseline_covs}{A character vector containing the column names of baseline covariates used in computing propensity scores.}

\item{trunc_weights}{A logical value. If TRUE, the computed weights are truncated to 1. Default is TRUE.}

\item{weight_threshold}{A numeric value. If set, filters out rows where the weight is below this threshold. Default is NULL.}
}
\value{
A dataframe similar to the input `data` but with additional columns:
\itemize{
  \item \strong{c_flag}: A binary flag indicating the current cohort versus the target cohort.
  \item \strong{ps_weight}: The computed propensity score for each observation.
  \item \strong{weight}: The computed cohort weight for each observation.
}
}
\description{
This function generates cohort weights for datasets with multiple cohorts. It utilizes propensity scores and the C-index to compute the weights.
}
\details{
The function begins by identifying the cohorts in the `data` and renaming the cohort column to "cohort".
It then computes the propensity scores for each cohort relative to the target cohort.
Following that, the function computes the inverse C-index using the propensity scores and cohort flags.
Finally, the cohort weights are calculated and either truncated or not based on the `trunc_weights` parameter.
Additionally, the function can filter out rows where the weight is below a specified threshold, if `weight_threshold` is set.
}
\examples{
# Generate a mock dataset
set.seed(123)
data <- data.frame(cohort_id = sample(1:3, 100, replace = TRUE),
                   covariate1 = runif(100),
                   covariate2 = rnorm(100))

# Use the create_cow function
weighted_data <- create_cow(data = data,
                           cohort_id = "cohort_id",
                           target_cohort = 2,
                           baseline_covs = c("covariate1", "covariate2"),
                           weight_threshold = 0.5)

}
