% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdf.R
\name{plot_cdfs}
\alias{plot_cdfs}
\title{Plot Cumulative Distribution Functions (CDFs)}
\usage{
plot_cdfs(data, variable, target_var, target, weight_var)
}
\arguments{
\item{data}{A dataframe containing the data to be visualized.}

\item{variable}{A character string indicating the column name of the continuous variable to visualize.}

\item{target_var}{A character string indicating the column name that differentiates cohorts.}

\item{target}{The value in the `target_var` column that defines the target cohort.}

\item{weight_var}{A character string indicating the column name for the weights.}
}
\value{
A `ggplot` object representing the CDFs.
}
\description{
This function plots the CDFs for a given variable in a dataset, and showcases the distributions
of the target cohort, the overall cohort, and a weighted global cohort.
}
\details{
The function produces three lines on the CDF plot:
\itemize{
  \item Target cohort CDF, colored in "#FF0155".
  \item Overall cohort CDF, colored in "#617989".
  \item Weighted global cohort CDF, colored in "#9447FB".
}
The function uses weighted empirical CDF calculations (`wtd.Ecdf`) and visualizes the results using `ggplot2`.
}
\examples{
# Generate a mock dataset
set.seed(123)
data <- data.frame(variable = rnorm(100),
                   target_var = sample(1:3, 100, replace = TRUE),
                   weight_var = runif(100))

# Use the plot_cdfs function
p <- plot_cdfs(data = data,
               variable = "variable",
               target_var = "target_var",
               target = 2,
               weight_var = "weight_var")
print(p)

}
